fixing some bugs, cst addition on LS and bundels and avoid loop unrolling

+++ src/s1p.c
@@ -40,19 +40,19 @@ typedef enum {
 
 static void compress_block(unsigned char *state, unsigned char *out,
                            const unsigned char *d, compress_mode mode,
-                           unsigned long long offset, unsigned long long n);
+                            unsigned int offset,  unsigned int n);
 
-static unsigned long long compress_data(shadow_state state,
+static  unsigned int compress_data(shadow_state state,
                                         unsigned char *out,
                                         const unsigned char *d,
-                                        unsigned long long dlen,
+                                         unsigned int dlen,
                                         compress_mode mode);
 
 static void init_sponge_state(shadow_state state,
                               const unsigned char *k, const unsigned char *p,
                               const unsigned char *n);
 static void xor_bytes(unsigned char* dest, const unsigned char* src1,
-               const unsigned char* src2, unsigned long long n);
+               const unsigned char* src2,  unsigned int n);
 
 void init_keys(const unsigned char **k, unsigned char p[P_NBYTES],
                const unsigned char *k_glob) {
@@ -80,9 +80,9 @@ static void init_sponge_state(shadow_state state,
   shadow(state);
 }
 
-void s1p_encrypt(unsigned char *c, unsigned long long *clen,
-                 const unsigned char *ad, unsigned long long adlen,
-                 const unsigned char *m, unsigned long long mlen,
+void s1p_encrypt(unsigned char *c,  unsigned int *clen,
+                 const unsigned char *ad,  unsigned int adlen,
+                 const unsigned char *m,  unsigned int mlen,
                  const unsigned char *k, const unsigned char *p,
                  const unsigned char *n) {
   // permutation state
@@ -93,7 +93,7 @@ void s1p_encrypt(unsigned char *c, unsigned long long *clen,
   compress_data(state, NULL, ad, adlen, AD);
 
   // compress message
-  unsigned long long c_bytes = 0;
+   unsigned int c_bytes = 0;
   if (mlen > 0) {
     state[RATE_BUNDLES][0] ^= 0x01;
     c_bytes = compress_data(state, c, m, mlen, PLAINTEXT);
@@ -106,9 +106,9 @@ void s1p_encrypt(unsigned char *c, unsigned long long *clen,
   *clen = c_bytes + CLYDE128_NBYTES;
 }
 
-int s1p_decrypt(unsigned char *m, unsigned long long *mlen,
-                const unsigned char *ad, unsigned long long adlen,
-                const unsigned char *c, unsigned long long clen,
+int s1p_decrypt(unsigned char *m,  unsigned int *mlen,
+                const unsigned char *ad,  unsigned int adlen,
+                const unsigned char *c,  unsigned int clen,
                 const unsigned char *k, const unsigned char *p,
                 const unsigned char *n) {
   // permutation state
@@ -119,7 +119,7 @@ int s1p_decrypt(unsigned char *m, unsigned long long *mlen,
   compress_data(state, NULL, ad, adlen, AD);
 
   // compress message
-  unsigned long long m_bytes = 0;
+   unsigned int m_bytes = 0;
   if (clen > CLYDE128_NBYTES) {
     state[RATE_BUNDLES][0] ^= 0x01;
     m_bytes = compress_data(state, m, c, clen - CLYDE128_NBYTES, CIPHERTEXT);
@@ -151,7 +151,7 @@ int s1p_decrypt(unsigned char *m, unsigned long long *mlen,
 // Only the XOR operation is performed, not XORing of padding constants.
 static void compress_block(unsigned char *state, unsigned char *out,
                            const unsigned char *d, compress_mode mode,
-                           unsigned long long offset, unsigned long long n) {
+                            unsigned int offset,  unsigned int n) {
   if (mode == CIPHERTEXT) {
     xor_bytes(out + offset, state, d + offset, n);
     memcpy(state, d + offset, n);
@@ -167,12 +167,12 @@ static void compress_block(unsigned char *state, unsigned char *out,
 // Input data buffer is d with length dlen.
 // Output is written into buffer out if mode is PLAINTEXT or CIPHERTEXT.
 // Padding is handled if needed.
-static unsigned long long compress_data(shadow_state state,
+static  unsigned int compress_data(shadow_state state,
                                         unsigned char *out,
                                         const unsigned char *d,
-                                        unsigned long long dlen,
+                                         unsigned int dlen,
                                         compress_mode mode) {
-  unsigned long long i;
+   unsigned int i;
   for (i = 0; i < dlen / RATE_BYTES; i++) {
     compress_block((uint8_t *)state, out, d, mode, i * RATE_BYTES, RATE_BYTES);
     shadow(state);
@@ -189,8 +189,8 @@ static unsigned long long compress_data(shadow_state state,
 
 // XOR buffers src1 and src2 into buffer dest (all buffers contain n bytes).
 void xor_bytes(unsigned char* dest, const unsigned char* src1,
-               const unsigned char* src2, unsigned long long n) {
-  for (unsigned long long i = 0; i < n; i++) {
+               const unsigned char* src2,  unsigned int n) {
+  for ( unsigned int i = 0; i < n; i++) {
     dest[i] = src1[i] ^ src2[i];
   }
 }
